Add files via upload
Updated function files. Now contains a function that can find if a note or set of notes is above a threshold or not within an audio clip and how long this note lasts for when STFT information is passed to it. This could then be used to create timing diagrams perhaps if some frequency was being seen to occur at too high a amplitude.